Converted bare constants in zscreen.py to named constants.
Also changed the calling convention of ZScreen.set_text_style() to take a sequence of style numbers rather than a single packed number--this will require a little more work on the end of the Z-Machine interpreter, but means less work (less bit-twiddling) and less duplicated code for the various ZScreen implementations.

(If such extra work really hampers the readability of the interpreter code, we can create separate adapter classes that do it.)

The basic prototypes for the ZScreen class and the other ZUI classes are mostly done now, I think, but I will probably be making further interface tweaks (such as the one I mentioned above) to make ZUI subclass implementation as simple and straightforward as possible.

My next step will be to finish a basic ZScreen implementation for trivialzui; this should be all we need, on the UI side of things, to make games playable.